+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-13 17:40:28
-// Commit: style(network): adjust design tokens, typography, and contrast in network
+// Updated at: 2026-08-14 15:28:07
+// Commit: feat(network): integrate custom telemetry logger in network
 
 class networkComponentState {
-  final String commitRef = '3588';
-  final String lastMessage = 'style(network): adjust design tokens, typography, and contrast in network';
-  final DateTime timestamp = DateTime.parse('2026-08-13T17:40:28+05:00');
+  final String commitRef = '3609';
+  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
+  final DateTime timestamp = DateTime.parse('2026-08-14T15:28:07+05:00');
 }